+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Reset password in admin-panel [#86](https://github.com/sham-ui/sham-ui-cli/issues/86)
 * Add page title in admin-panel [#87](https://github.com/sham-ui/sham-ui-cli/issues/87)
 * Refactoring server code in admin-panel [#88](https://github.com/sham-ui/sham-ui-cli/issues/88)
+* Add confirmation modal on edit name/email/password in admin-panel
 
 
 ## 4.3.0